pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "triple-c-rag"
version = "0.1.0"
description = "RAG (Retrieval-Augmented Generation) Test Module"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "지원"},
    {name = "지윤"},
    {name = "동훈"},
]
keywords = ["rag", "retrieval", "augmented", "generation", "llm", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "openai>=1.0.0",
    "chromadb>=0.4.0",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]
all = [
    "PyPDF2>=3.0.0",
    "python-docx>=0.8.11",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "pinecone-client>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/your-repo/triple-c-rag"
Repository = "https://github.com/your-repo/triple-c-rag"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | data
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["src", "config"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
